Support for postable_by
Add support for postable_by on boards and categories. This involved
updating all the write based auth methods to include checking whether
the user's priority matched or was greater than the priority
(postable_by) of the board, parent boards, and parent category. All the
auth methods were also refactored because the auth methods themselves
still retained their function names.

A new DB method getBoardWriteAccess was created and
getThreadBoardInBoardMapping was refactored to handle a board depth
lower than 2.

The byBoard route was updated to return a new property, 'writeAccess',
that takes the user's priority and checks it against the postable_by of
the current board and all of it's parents to see if the user can do a
write based action against this board.