Bug Fix: edit thread title
There was a regression that lead to an error whenever a thread's title
was edited. This regression happened while trying to add support for
postable_by for boards. The support was added but the title's auth
method relying on sending back the first post for the thread as well.
The new postable_by condition shifted the index of the first post data,
but the index value in the .then was not updated.