1.0.4: Hooks system on threads.create, byBoards order by
The hooks system was added to the threads.create route. The primary
reason for this was to add the auto-moderation function to this route.

The byBoard db method was also updated to move the order by sorting to
the end of the query rather than the start. This guarantees that the
ordering will always be correct.